Fix segfault with the open puzzle solution dialog on Android
Closes #171

+++ src/ui_android.cc
@@ -259,7 +259,7 @@ extern "C" void
 Java_com_bithack_principia_PrincipiaBackend_addActionAsInt(JNIEnv *env, jclass jcls,
         jint action_id, jlong action_data)
 {
-    uint32_t d = (uint32_t)((int64_t)action_data);
+    uint64_t d = (uint64_t)((int64_t)action_data);
     P.add_action(action_id, d);
 }
